+++ simple-stack/src/main/java/com/zhuinden/simplestack/BackstackManager.java
@@ -160,6 +160,10 @@ public class BackstackManager
     Map<Object, SavedState> keyStateMap = new HashMap<>();
     ScopeManager scopeManager = new ScopeManager();
 
+    /* init */ {
+        scopeManager.setBackstackManager(this);
+    }
+
     StateChanger stateChanger;
 
     /**